requirements.txt
asgiref==3.5.2
attrs==21.4.0
awscli==1.25.81
beautifulsoup4==4.11.1
boto3==1.24.80
botocore==1.27.80
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.4
cryptography==38.0.1
defusedxml==0.7.1
dj-rest-auth==2.1.12
Django==4.0.6
django-allauth==0.51.0
django-cors-headers==3.13.0
django-filter==21.1
djangorestframework==3.13.0
djangorestframework-simplejwt==5.0.0
docutils==0.16
drf-spectacular==0.23.1
drf-spectacular-sidecar==2022.7.1
drf-writable-nested==0.6.4
idna==3.4
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.7.2
Markdown==3.3.7
mysqlclient==2.1.1
numpy==1.23.4
oauthlib==3.2.2
pandas==1.4.3
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.18.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==5.4.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.7.2
s3transfer==0.6.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.12
